+++ ProperSave/ProperSavePlugin.cs
@@ -25,7 +25,7 @@ namespace ProperSave
     {
         public const string GUID = "com.KingEnderBrine.ProperSave";
         public const string Name = "Proper Save";
-        public const string Version = "3.0.5";
+        public const string Version = "3.0.6";
 
         private static readonly char[] invalidSubDirectoryCharacters = new[] { '\\', '/', '.' };